+++ src/ska_oso_services/engineering/api/ebs.py
@@ -1,9 +1,9 @@
 import logging
 
 from fastapi import APIRouter
+from ska_db_oda.common.uow import UnitOfWork
 from ska_db_oda.postgres.mapping import StatusLabel
 from ska_db_oda.repository.status import Status
-from ska_db_oda.rest.fastapicontext import UnitOfWork
 from ska_oso_pdm import OSOExecutionBlock as ExecutionBlock
 from ska_oso_pdm import TelescopeType
 from ska_oso_pdm.execution_block import RequestResponse
@@ -82,7 +82,7 @@ def set_eb_status_observed(eb_id: str, oda: UnitOfWork) -> Status:
 def set_eb_status_failed(eb_id: str, oda: UnitOfWork) -> Status:
     with oda as uow:
         uow.status.update_status(
-            entity_id=eb_id, status=StatusLabel.FAILED, updated_by=NOTEBOOK_USER
+            entity_id=eb_id, status=StatusLabel.OBSERVING_FAILED, updated_by=NOTEBOOK_USER
         )
         uow.commit()
         return uow.status.get_current_status(entity_id=eb_id)