+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 14.5.0
 ***********
+* [BREAKING] New major version of ODA Updated to ODA v18.0.0
 * [BUGFIX] Fixes error message for MID targets below the elevation range
 *  Improves validation of Solar System objects
 * [BUGFIX] Restore LOW targets below 45 degree elevation warning